feat: add versioning routes and handler exception
