+++ app/Providers/RouteServiceProvider.php
@@ -30,11 +30,12 @@ class RouteServiceProvider extends ServiceProvider
 
         $this->routes(function () {
             Route::middleware('api')
-                ->prefix('api')
-                ->group(base_path('routes/api.php'));
+                ->prefix('api/v1')
+                ->group(base_path('routes/v1/no-auth.api.php'));
 
-            Route::middleware('web')
-                ->group(base_path('routes/web.php'));
+            Route::middleware(['api', 'jwt.verify'])
+                ->prefix('api/v1')
+                ->group(base_path('routes/v1/auth.api.php'));
         });
     }
 }
